pyproject.toml
[project]
name = "eodata-gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "brustt", email = "martin.dzr@protonmail.com" }
]
requires-python = ">=3.8"
dependencies = [
    "eodag>=3.3.1",
    "ipykernel>=6.29.5",
    "openeo>=0.40.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
eodata-gateway = "eodata_gateway:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
